Implement Enemy Ai (Complete) (Kind of)

+++ Assets/Scripts/Interfaces/ITakeDamage.cs
@@ -1,4 +1,4 @@
 public interface ITakeDamage
 {
-    void TakeDamage(float damage);
+    void TakeDamage(int damage);
 }
